for #894 use multithreading to access database for Statement
